IAR - exporter iar removal, logic moved to IARWorkbench class
